move to dist/, call it "bundle"
